Refactor common entityservice tests into a base class
