+++ metadata-io/src/main/java/com/linkedin/metadata/entity/AspectDao.java
@@ -1,0 +1,5 @@
+package com.linkedin.metadata.entity;
+
+public interface AspectDao {
+    EntityAspect getAspect(String urn, String aspectName, long version);
+}
