+++ metadata-io/src/main/java/com/linkedin/metadata/entity/EntityAspect.java
@@ -1,0 +1,5 @@
+package com.linkedin.metadata.entity;
+
+public interface EntityAspect {
+    String getSystemMetadata();
+}
