+++ metadata-io/src/main/java/com/linkedin/metadata/entity/ebean/EbeanAspectV2.java
@@ -1,5 +1,6 @@
 package com.linkedin.metadata.entity.ebean;
 
+import com.linkedin.metadata.entity.EntityAspect;
 import io.ebean.Model;
 import io.ebean.annotation.Index;
 import java.sql.Timestamp;
@@ -26,7 +27,7 @@ import lombok.Setter;
 @AllArgsConstructor
 @Entity
 @Table(name = "metadata_aspect_v2")
-public class EbeanAspectV2 extends Model {
+public class EbeanAspectV2 extends Model implements EntityAspect {
 
   private static final long serialVersionUID = 1L;
